[feat] Update the response in middlewares when there's unexpected response from model or service functions

+++ server/controller/ff/file_delete_controller.js
@@ -18,8 +18,10 @@ import {
 	deleteAllVersionsForOneObject,
 	deleteFolderAndContents,
 } from "../../service/s3/s3_delete.js";
+
+import { customError } from "../../error/custom_error.js";
 // ======================================================================
-const deleteDB = async (req, res) => {
+const deleteDB = async (req, res, next) => {
 	console.log("/deleteDB: ", req.body);
 	// req.body = { "delList": ["folder/", "file.ext"] }
 	const { delList, parentPath } = req.body;
@@ -29,7 +31,6 @@ const deleteDB = async (req, res) => {
 
 	for (let i = 0; i < delList.length; i++) {
 		const key = delList[i];
-
 		if (key.endsWith("/")) {
 			console.log("delete folder");
 
@@ -37,32 +38,48 @@ const deleteDB = async (req, res) => {
 			const folders = key.slice(0, key.length - 1).split("/");
 			const parentId = await iterForParentId(userId, folders);
 			console.log("parentId: ", parentId);
+      if (parentId === -1) {
+        return next(customError.badRequest("No such key"));
+      }
 			const deleteRecurRes = await deleteRecur(parentId, userId, nowTime);
 			console.log("deleteRecurRes: ", deleteRecurRes);
+      if (!deleteRecurRes) {
+        return next(customError.internalServerError());
+      }
 		} else {
 			console.log("delete file");
 
 			// DB
 			const fileId = await findFileIdByPath(userId, key);
 			console.log("fileId: ", fileId);
+      if (!fileId) {
+        return next(customError.badRequest("No such key"));
+      }
 			const deleteRes = await markDeleteById(nowTime, fileId);
 			console.log("deleteRes: ", deleteRes);
+      if (!deleteRes) {
+        return next(customError.internalServerError());
+      }
 		}
 	}
 
 	// update usage of an user
 	const currentUsed = await updateSpaceUsedByUser(userId, nowTime);
+  if (currentUsed === -1) {
+    return next(customError.internalServerError());
+  }
 	req.session.user.used = currentUsed;
 
 	// emit list
   const io = req.app.get("socketio");
 	emitNewList(io, userId, parentPath);
+  emitTrashList(io, userId);
   emitUsage(io, userId, req.session.user);
 
 	return res.send("ok");
 };
 
-const permDelete = async (req, res) => {
+const permDelete = async (req, res, next) => {
 	console.log("/perm-delete: ", req.body);
 	const { permDeleteList } = req.body;
 	const userId = req.session.user.id;
@@ -79,10 +96,16 @@ const permDelete = async (req, res) => {
 			const folders = key.slice(0, key.length - 1).split("/");
 			const parentId = await iterForParentId(userId, folders);
 			console.log("parentId: ", parentId);
+      if (parentId === -1) {
+        return next(customError.badRequest("No such key"));
+      }
 
 			// update DB
 			const deleteDB = await permDeleteRecur(parentId, userId);
 			console.log("deleteDB: ", deleteDB);
+      if (!deleteDB) {
+        return next(customError.internalServerError());
+      }
 
 			// update S3
 			const deleteS3 = await deleteFolderAndContents(
@@ -95,11 +118,16 @@ const permDelete = async (req, res) => {
 			// get fileId
 			const fileId = await findDeletedFileIdByPath(userId, key);
 			console.log("fileId: ", fileId);
-			// TODO: check first if this file is marked as deleted?
+			if (!fileId) {
+        return next(customError.badRequest("No such key"));
+      }
 
 			// update DB
 			const deleteDB = await permDeleteByFileId(fileId);
 			console.log("deleteDB: ", deleteDB);
+      if (!deleteDB) {
+        return next(customError.internalServerError());
+      }
 
 			// update S3
 			const deleteS3 = await deleteAllVersionsForOneObject(
